chore: move create condition object to consumer function
