+++ src/libecalc/domain/infrastructure/energy_components/legacy_consumer/tabulated/tabular_consumer_function.py
@@ -25,6 +25,7 @@ from libecalc.domain.infrastructure.energy_components.legacy_consumer.tabulated.
 )
 from libecalc.domain.process.core.results import EnergyFunctionResult
 from libecalc.expression import Expression
+from libecalc.expression.expression import ExpressionType
 
 
 class TabularConsumerFunction(ConsumerFunction):
@@ -56,7 +57,7 @@ class TabularConsumerFunction(ConsumerFunction):
         energy_usage_adjustment_constant: float,
         energy_usage_adjustment_factor: float,
         variables_expressions: list[VariableExpression],
-        condition: Condition,
+        condition_expression: ExpressionType | None = None,
         power_loss_factor_expression: Expression | None = None,
     ):
         """Tabulated consumer function [MW] (energy) or [Sm3/day] (fuel)."""
@@ -68,9 +69,9 @@ class TabularConsumerFunction(ConsumerFunction):
             energy_usage_adjustment_constant=energy_usage_adjustment_constant,
             energy_usage_adjustment_factor=energy_usage_adjustment_factor,
         )
+        self._condition_expression = condition_expression
         self._variables_expressions = variables_expressions
 
-        self.condition = condition
         # Typically used for power line loss subsea et.c.
         self._power_loss_factor_expression = power_loss_factor_expression
 
@@ -93,6 +94,10 @@ class TabularConsumerFunction(ConsumerFunction):
         """
 
         variables_for_calculation = []
+        condition = Condition(
+            expression_input=self._condition_expression,
+            expression_evaluator=expression_evaluator,
+        )
         # If some of these are rates, we need to calculate stream day rate for use
         # Also take a copy of the calendar day rate and stream day rate for input to result object
         for variable in self._variables_expressions:
@@ -110,12 +115,12 @@ class TabularConsumerFunction(ConsumerFunction):
 
         # for tabular, is_valid is based on energy_usage being NaN. This will also (correctly) change potential
         # invalid points to valid where the condition sets energy_usage to zero
-        energy_function_result.energy_usage = self.condition.apply_to_array_as_list(
+        energy_function_result.energy_usage = condition.apply_to_array_as_list(
             np.asarray(energy_function_result.energy_usage)
         )
 
         energy_function_result.power = (
-            self.condition.apply_to_array_as_list(np.asarray(energy_function_result.power))
+            condition.apply_to_array_as_list(np.asarray(energy_function_result.power))
             if energy_function_result.power is not None
             else None
         )
@@ -129,7 +134,7 @@ class TabularConsumerFunction(ConsumerFunction):
             periods=expression_evaluator.get_periods(),
             is_valid=np.asarray(energy_function_result.is_valid),
             energy_function_result=energy_function_result,
-            condition=self.condition.as_vector(),
+            condition=condition.as_vector(),
             energy_usage_before_power_loss_factor=np.asarray(energy_function_result.energy_usage),
             power_loss_factor=power_loss_factor,
             energy_usage=apply_power_loss_factor(
