+++ src/libecalc/presentation/yaml/mappers/component_mapper.py
@@ -201,7 +201,6 @@ class ConsumerMapper:
             energy_usage_model = self.__energy_usage_model_mapper.from_yaml_to_dto(
                 data.energy_usage_model,
                 consumes=consumes,
-                expression_evaluator=expression_evaluator,
             )
         except InvalidEnergyUsageModelException as e:
             energy_usage_model_yaml_path = yaml_path.append("ENERGY_USAGE_MODEL")
