+++ tests/conftest.py
@@ -12,8 +12,7 @@ from libecalc.common.energy_usage_type import EnergyUsageType
 from libecalc.common.math.numbers import Numbers
 from libecalc.common.time_utils import Frequency, Period, Periods
 from libecalc.common.utils.rates import RateType
-from libecalc.common.variables import VariablesMap, ExpressionEvaluator
-from libecalc.domain.condition import Condition
+from libecalc.common.variables import VariablesMap
 from libecalc.domain.infrastructure.energy_components.legacy_consumer.consumer_function.direct_expression_consumer_function import (
     DirectExpressionConsumerFunction,
 )
@@ -346,42 +345,3 @@ def with_neqsim_service():
     neqsim_service = NeqsimService()
     yield neqsim_service
     neqsim_service.shutdown()
-
-
-@pytest.fixture
-def direct_expression_model_factory(condition_factory):
-    def create_direct_expression_model(
-        expression: Expression,
-        energy_usage_type: EnergyUsageType,
-        consumption_rate_type: RateType = RateType.STREAM_DAY,
-    ):
-        if energy_usage_type == EnergyUsageType.POWER:
-            return DirectExpressionConsumerFunction(
-                energy_usage_type=energy_usage_type,
-                condition=condition_factory(),
-                load=expression,
-                power_loss_factor=None,
-                consumption_rate_type=consumption_rate_type,
-            )
-        else:
-            return DirectExpressionConsumerFunction(
-                energy_usage_type=energy_usage_type,
-                condition=condition_factory(),
-                fuel_rate=expression,
-                power_loss_factor=None,
-                consumption_rate_type=consumption_rate_type,
-            )
-
-    return create_direct_expression_model
-
-
-@pytest.fixture
-def condition_factory(expression_evaluator_factory):
-    def create_condition(
-        expression_input: Expression | None = None, expression_evaluator: ExpressionEvaluator = None
-    ) -> Condition | None:
-        if expression_evaluator is None:
-            expression_evaluator = expression_evaluator_factory.default()
-        return Condition(expression_input=expression_input, expression_evaluator=expression_evaluator)
-
-    return create_condition
